[GC] 8209376: ZGC: Move ZMarkStackAllocator into a separate file
Summary: Backport JDK-8209376 for ZGC. (https://github.com/openjdk/jdk/commit/a51aeedca0)


Test Plan: test/hotspot/jtreg/gc

Issue: https://github.com/alibaba/dragonwell11/issues/60
